+++ __mocks__/game.ts
@@ -142,6 +142,9 @@ const game = {
     },
     world: {
         id: 'worldId'
+    },
+    release: {
+        generation: 11
     }
 };